requirements.txt
africastalking==1.1.7
appdirs==1.4.3
attrs==19.3.0
awscli==1.19.0
bcrypt==3.2.0
blessed==1.15.0
botocore==1.20.0
cached-property==1.5.1
cement==2.8.2
certifi==2019.6.16
cffi==1.14.3
chardet==3.0.4
colorama==0.3.9
contextlib2==0.5.5
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-appconf==1.0.3
django-axes==4.5.4
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-ipware==2.1.0
django-rest-auth==0.9.5
django-session-timeout==0.0.4
django-storages==1.7.1
django-stubs==1.1.0
djangorestframework==3.9.4
djangorestframework-api-key==0.3.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djangorestframework-stubs==1.0.0
docker==3.7.3
docker-compose==1.25.5
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
gunicorn==19.9.0
idna==2.7
isodate==0.6.0
jmespath==0.9.4
jsonschema==2.6.0
M2Crypto==0.37.1
mypy==0.720
mypy-extensions==0.4.3
parameterized==0.8.1
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
pisa==3.0.33
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.4
schema==0.7.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.1.7
websocket-client==0.56.0
